chore: update crud trip

+++ src/services/trip.service.ts
@@ -5,6 +5,10 @@ import TripRepository from '../repositories/trip.repository';
 import Security from 'utility-layer/dist/security'
 import { Service } from 'fastify-decorators';
 import { In } from 'typeorm';
+import BankAccountRepository from '../repositories/bank-account.reepository';
+import PaymentShipperRepository from '../repositories/payment-shipper.repsitory';
+import PaymentCarrierRepository from '../repositories/payment-carrier.repository';
+import { PaymentShipper, PaymentCarrier, Trip } from '../models';
 
 interface FindTripProps {
   descending?: boolean
@@ -18,11 +22,28 @@ interface ITruckProps {
   startDate: string
 }
 
+export interface IShipmentTrip {
+  shipperPricePerTon?: number
+  shipperPaymentStatus?: "PAYMENT_DUE" | "PAID" | "VOID"
+  shipperBillStartDate?: string
+  shipperPaymentDate?: string
+
+  weightStart?: number
+  weightEnd?: number
+  carrierPricePerTon?: number
+  bankAccountId?: number
+  carrierPaymentStatus?: "PAID" | "AWAITING" | "APPROVED" | "REJECTED" | "ISSUED"
+  carrierPaymentDate?: string
+}
+
 const security = new Security();
 const vwTripInprogressRepository = new VwTripInprogressRepository();
 const truckRepository = new TruckRepository();
 const jobCarrierRepository = new JobCarrierRepository();
 const tripRepository = new TripRepository();
+const bankAccRepository = new BankAccountRepository();
+const paymentShipperRepository = new PaymentShipperRepository();
+const paymentCarrierRepository = new PaymentCarrierRepository();
 
 const camelToSnakeCase = (str: string) => str.replace(/[A-Z]/g, letter => `_${letter.toLowerCase()}`);
 
@@ -35,6 +56,13 @@ const diffArray = (arr1: Array<any>, arr2: Array<any>): Array<any> => {
   const diff2 = diff(arr2, arr1);
   return [...diff1, ...diff2];
 }
+
+const encodeIds = (data: any[]) => data?.map((attr: any) => ({
+  ...attr,
+  id: security.encodeUserId(attr.id)
+})) ?? [];
+
+const encodeId = (data: any) => data ? { ...data, id: security.encodeUserId(data.id) } : null;
 
 @Service()
 export default class TripService {
@@ -94,18 +122,86 @@ export default class TripService {
     }
   }
 
+  async getTripDetail(tripId: string): Promise<any> {
+    const decodeTripId = security.decodeUserId(tripId);
+    const jobCarrier = await jobCarrierRepository.getJobAndTruckIdByJobCarrierId(decodeTripId);
+    if (!jobCarrier?.length) {
+      throw new Error('Request not found');
+    }
+    console.log('jobCarrier[0] :>> ', jobCarrier[0]);
+    const {
+      job_id: jobId,
+      truck_id: truckId,
+      weight_start: weightStart,
+      weight_end: weightEnd
+    } = jobCarrier[0];
+    const jobDetail = await tripRepository.getJobDetail(jobId);
+    const truckDetail = await tripRepository.getTruckDetail(truckId);
+    jobDetail.id = security.encodeUserId(jobDetail.id);
+    if (jobDetail?.owner?.id) {
+      jobDetail.owner.id = security.encodeUserId(jobDetail.owner.id)
+    }
+
+    truckDetail.id = security.encodeUserId(truckDetail.id);
+    if (truckDetail?.owner?.id) {
+      truckDetail.owner.id = security.encodeUserId(truckDetail.owner.id)
+    }
+
+    console.log('jobDetail :>> ', jobDetail);
+    console.log(`truckDetail`, truckDetail);
+
+    // get bank account
+    const bankAccounts = await bankAccRepository.find({
+      select: ['id', 'accountName', 'accountNo', 'bankName'],
+      where: { userId: truckDetail.carrierId }
+    });
+
+    const bankAccountDecrypted = encodeIds(bankAccounts);
+
+    // get payment shipper
+    const paymentShipper = await paymentShipperRepository.findByTripId(decodeTripId, {
+      select: ['id', 'pricePerTon', 'amount', 'feeAmount', 'feePercentage', 'netAmount', 'paymentStatus', 'billStartDate', 'paymentDate']
+    });
+    const paymentShipperDecrypted = encodeId(paymentShipper);
+
+    // get payment carrier
+    const paymentCarrier = await paymentCarrierRepository.findByTripId(decodeTripId, {
+      select: ['id', 'pricePerTon', 'bankAccountId', 'amount', 'feeAmount', 'feePercentage', 'netAmount', 'paymentStatus', 'paymentDate']
+    });
+    const paymentCarrierDecrypted = encodeId(paymentCarrier);
+
+    return {
+      id: tripId,
+      weightStart,
+      weightEnd,
+      bankAccount: bankAccountDecrypted,
+      job: {
+        ...jobDetail,
+        payment: paymentShipperDecrypted
+      },
+      truck: {
+        ...truckDetail,
+        carrierId: security.encodeUserId(truckDetail.carrierId),
+        payment: paymentCarrierDecrypted
+      },
+    };
+  }
+
   async bulkTrip(jobId: string, truckIds: ITruckProps[], userId: number): Promise<any> {
     const truckDate: any = {};
     const decJobId = security.decodeUserId(jobId);
     const decTruckIds = truckIds.map((truck: ITruckProps) => {
-      truckDate[truck.id] = truck.startDate;
-      return security.decodeUserId(truck.id)
+      const decTruckId = security.decodeUserId(truck.id);
+      truckDate[decTruckId] = truck.startDate;
+      return decTruckId;
     });
     const trucks = await truckRepository.findManyById(decTruckIds);
     const carrierIds = trucks.map((truck: any) => truck.carrier_id);
     console.log('decJobId :>> ', decJobId);
     console.log('trucks :>> ', trucks);
     console.log('carrierIds :>> ', carrierIds);
+    console.log('decTruckIds :>> ', decTruckIds);
+    console.log('truckDate :>> ', truckDate);
 
     const uniqueCarrierIds = [...new Set(carrierIds)];
     console.log('uniqueCarrierIds :>> ', uniqueCarrierIds);
@@ -116,7 +212,7 @@ export default class TripService {
     const result = await Promise.all(trucks.map(async (truck: any) => {
       const jobCarrier = jobCarriers.find((jobCarrier: any) => jobCarrier.carrierId === truck.carrier_id);
       console.log('find jobCarrier :>> ', jobCarrier);
-      return tripRepository.add({
+      const tripData = await tripRepository.add({
         jobCarrierId: jobCarrier.id,
         truckId: truck.id,
         priceType: null,
@@ -124,7 +220,15 @@ export default class TripService {
         updatedAt: new Date(),
         createdUser: userId.toString(),
         startDate: truckDate[truck.id]
-      })
+      });
+
+      // initial payment shipper
+      await paymentShipperRepository.add({ tripId: tripData.id, feePercentage: '1' });
+
+      // initial payment carrier
+      await paymentCarrierRepository.add({ tripId: tripData.id, feePercentage: '1' });
+
+      return tripData;
     }))
 
     console.log('result :>> ', result);
@@ -213,9 +317,89 @@ export default class TripService {
     );
   }
 
-  async deleteTripById(tripId: string): Promise<void> {
+  async deleteTripById(tripId: string, userId: number): Promise<void> {
     const decTripId = security.decodeUserId(tripId);
-    await tripRepository.delete(decTripId);
+    await tripRepository.delete(decTripId, { updatedUser: userId.toString() });
+  }
+
+  async updateShipmentTrip({ tripId, userId, data }: { tripId: string; userId: string; data: IShipmentTrip; }): Promise<void> {
+    const decodeTripId = security.decodeUserId(tripId);
+    const decodeUserId = security.decodeUserId(userId);
+
+    // get payment shipper
+    const paymentShipper = await paymentShipperRepository.findByTripId(decodeTripId, { select: ['feePercentage'] });
+    const shipperFeePercentage = paymentShipper?.feePercentage ?? 1;
+
+    // get payment carrier
+    const paymentCarrier = await paymentCarrierRepository.findByTripId(decodeTripId, { select: ['feePercentage'] });
+    const carrierFeePercentage = paymentCarrier?.feePercentage ?? 1;
+
+    // update payment
+    let paymentShipperData: Partial<PaymentShipper> = {
+      updatedAt: new Date(),
+      updatedUser: decodeUserId
+    };
+    let paymentCarrierData: Partial<PaymentCarrier> = {
+      updatedAt: new Date(),
+      updatedUser: decodeUserId
+    };
+
+    if (data?.shipperPricePerTon) {
+      const amount = data.weightEnd ? (data.weightEnd * data.shipperPricePerTon) : 0;
+      const feeAmount = amount * (+shipperFeePercentage / 100);
+      const netAmount = Math.abs(amount - feeAmount);
+
+      paymentShipperData.pricePerTon = data.shipperPricePerTon.toString();
+      paymentShipperData.amount = amount.toString();
+      paymentShipperData.feeAmount = feeAmount.toString();
+      paymentShipperData.netAmount = netAmount.toString();
+    }
+
+    paymentShipperData = {
+      ...paymentShipperData,
+      ...(!paymentShipper ? { tripId: decodeTripId, createdAt: new Date() } : undefined),
+      ...(data?.shipperPaymentStatus ? { paymentStatus: data.shipperPaymentStatus } : undefined),
+      ...(data?.shipperBillStartDate ? { billStartDate: data.shipperBillStartDate } : undefined),
+      ...(data?.shipperPaymentDate ? { paymentDate: data.shipperPaymentDate } : undefined)
+    }
+
+    await paymentShipperRepository.updateByTripId(decodeTripId, paymentShipperData);
+
+    if (data?.carrierPricePerTon) {
+      const amount = data.weightEnd ? (data.weightEnd * data.carrierPricePerTon) : 0;
+      const feeAmount = amount * (+carrierFeePercentage / 100);
+      const netAmount = Math.abs(amount - feeAmount);
+
+      paymentCarrierData.pricePerTon = data.carrierPricePerTon.toString();
+      paymentCarrierData.amount = amount.toString();
+      paymentCarrierData.feeAmount = feeAmount.toString();
+      paymentCarrierData.netAmount = netAmount.toString();
+    }
+
+    paymentCarrierData = {
+      ...paymentCarrierData,
+      ...(!paymentCarrier ? { tripId: decodeTripId, createdAt: new Date() } : undefined),
+      ...(data?.bankAccountId ? { bankAccountId: data.bankAccountId } : undefined),
+      ...(data?.carrierPaymentDate ? { paymentDate: data.carrierPaymentDate } : undefined),
+      ...(data?.carrierPaymentStatus ? { paymentStatus: data.carrierPaymentStatus } : undefined)
+    }
+
+    await paymentCarrierRepository.updateByTripId(decodeTripId, paymentCarrierData);
+
+    let tripData: Partial<Trip> = {
+      updatedAt: new Date(),
+      updatedUser: decodeUserId
+    };
+
+    if (data?.weightStart) {
+      tripData.weightStart = data.weightStart.toString();
+    }
+
+    if (data?.weightEnd) {
+      tripData.weightEnd = data.weightEnd.toString();
+    }
+
+    await tripRepository.update(decodeTripId, tripData);
   }
 
 }
